--- Face recognition/.idea/inspectionProfiles/Project_Default.xml
@@ -1,13 +1,0 @@
-<component name="InspectionProjectProfileManager">
-  <profile version="1.0">
-    <option name="myName" value="Project Default" />
-    <inspection_tool class="Eslint" enabled="true" level="WARNING" enabled_by_default="true" />
-    <inspection_tool class="PyUnresolvedReferencesInspection" enabled="true" level="WARNING" enabled_by_default="true">
-      <option name="ignoredIdentifiers">
-        <list>
-          <option value="pulp" />
-        </list>
-      </option>
-    </inspection_tool>
-  </profile>
-</component>
