--- Face recognition/encryption_key.key
@@ -1,1 +1,0 @@
-8b95WoeL0lFTH_O8LAJjt_luaEhTfIGjIMZWKDuA1Nw=
